+++ src/MOI_wrapper/results.jl
@@ -3,16 +3,16 @@ function MOI.get(model::Optimizer, ::MOI.TerminationStatus)
 end
 
 function MOI.get(model::Optimizer, ::MOI.ObjectiveValue)
-    # if model.status == MOI.OPTIMIZE_NOT_CALLED
-    #     @error "optimize! not called"
-    # end
+    if model.status == MOI.OPTIMIZE_NOT_CALLED
+        @error "optimize! not called"
+    end
     return model.inner.best_sol
 end
 
 function MOI.get(model::Optimizer, ::MOI.ObjectiveBound)
-    # if model.status == MOI.OPTIMIZE_NOT_CALLED
-    #     @error "optimize! not called"
-    # end
+    if model.status == MOI.OPTIMIZE_NOT_CALLED
+        @error "optimize! not called"
+    end
     return model.inner.best_bound
 end
